% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Implementation of the Euclidean Algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A numeric scalar representing the first number.}

\item{b}{A numeric scalar representing the second number.}
}
\value{
The greatest common divisor of \code{a} and \code{b} as a numeric
    scalar.
}
\description{
The function \code{euclidean} finds the greatest common divisor
    of two numbers using the euclidean algorithm.
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)


}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
